Visitor: fix argument order in with_locals
with_locals is meant to surround all side-effects of a function call
with calls to the enter_scope and leave_scope methods.
To achieve this, it needs to be applied to a function that only
needs one more function argument.
This fixes a bug where almost all uses of with_locals involved a
function that required two more arguments.

This fix changes the interface to make it easier to do the
right thing.