Number tmp files
To make it easier to understand the sequence of transformations,
this change adds numbers to the filenames.

The numbers are multiples of 10 so that it is easy to insert new
transformations without having to change other lines.

(At some point, we will probably create a better framework for
this that automatically assigns numbers, makes it easy to
enable/disable/reorder transforms, etc. But this is an easy change
that fixes the immediate need and does not make it harder to
adopt something better when we have time to write it.)