Cleanup: AST manipulation helper functions
This change moves some of the AST manipulation functions
that were scattered around the codebase to asl_utils.ml.

As we add new transformations, we are slowly adding more
and more of these functions so this is a good time to
collect them all together so that there is a clear place
where they functions should be defined.

This is probably not a complete list of all the helper functions that we
will want but it includes the ones that we have at the moment or that
are useful in transformations that are under development.