xform_simplify_expr: new transform
This defines a transformation on expressions that simplifies
expressions such as

    x - x   ==>  0
    2*x + x ==>  3*x

This version is fairly restricted in the optimizations that
it performs. As we tackle more complex ASL code, there is scope
for making this optimize more cases but I want to start with
a simple version and make it more complex only as we see
the need.