desugaring: don't desugar slices in tcheck
This change reduces the amount of desugaring done in the
typechecker. In particular, it prevents Expr_Slices from
being desugared to asl_extract_bits.

The typechecker has been desugaring some bitslice constructs
to make code generation easier.
Now that we have a lot more transformations that can handle bitslices,
this is not helpful and actually complicates things because
it prevented some of the later transformations from recognizing
bitslice operations.