Add let expressions to the AST
This adds let expressions to the AST so that it can be used
in transformations. (It is not intended for use in specifications
so the new keywords begin with double-underscore so that we can easily
detect if it does get used.)

The expression "__let x : ty = expr1 __in expr2" means the same as it
does in OCaml: evaluate 'expr1', temporarily bind 'x' to the resulting
value and then evaluate 'expr2' in that context.
It is especially useful in transformations that need to refer to the
value of 'x' multiple times and where it is possible that 'expr1' might
have side effects, throw an exception or fail an assertion or range
check.

The motivation for this is to make it easier to write transformations
without having to worry about whether 'expr1' has side effects or
throws exceptions and whether the expression is conditionally executed.
For example, the following transformation would not be valid (because
the subscript 'XMM[i]' could be evaluated even if 'i' is out of range.

    let p = if i IN {0..31} then XMM[i][31:0] else Zeros(128);
    ==>
    let __tmp = XMM[i];
    let p = if i IN {0..31} then __tmp[31:0] else Zeros(128);

Using let-expressions, the same transformation can be written

    let p = if i IN {0..31} then XMM[i][31:0] else Zeros(128);
    ==>
    let p = if i IN {0..31} then (__let __tmp = XMM[i] __in __tmp[31:0])
                            else Zeros(128);

The C backend can support this directly using the statement-expression
syntax (https://gcc.gnu.org/onlinedocs/gcc/Statement-Exprs.html)
supported by both gcc and clang.

    ({ ty x = expr1; expr2 })

Since Verilog does not support statement expressions, the Verilog
backend would require a lowering pass that converts let-expressions
to let-statements. For example:

    var p : bits(32);
    if i IN {0..31} then
        let __tmp = XMM[i];
        p = __tmp[31:0];
    else
        p = Zeros(128);
    end

As this example shows, this transformation still has to be careful about
conditional expressions but, it will be easier to get this single
transformation right than to remember to deal correctly with conditional
expressions in all the transformations that let-bind an expression.