Split generated _funs.c into several
Add the option --num-c-files specifying the number of .c files produced.
It is intended to shorten compile times which tend to be rather long if
much C code is collected in the same file. Default is 1.

For now, apply the option to _funs.c only.

The number of .c files is approximately the total number of declarations
divided by the --num-c-files.

Each .c file gets _N suffix appended to the filename, e.g. funs_5.c