mono: Cache ident to declaration mapping
Do the same thing in xform_rmw for consistency.

Before this commit, find_decl compared identifiers using shallow
comparisons and not using Ident.compare. But this function is removed
anyway.

I tried caching identifiers using both a Map and a Hashtbl. Hashtbl was
about 0.5 seconds faster when compiling the ia-spec. The downside with
using a Hashtbl is that it will use more memory. Since our hash function
is generating unique hashes, the table will be as big as the biggest
hash id. A Map would only store the actual ids that we use.

I did the same experiment with storing idents in idents.ml. But Map was
faster for IdentInfo.t.