Automatically introduce case splits
This extends the mono phase to expand a polymorphic call site to a
Stmt_Case when exact values of the parameters to the call site are
missing but constraints are present. The children of the Stmt_Case
containing calls will be monomorphized using constant propagation.

A binding operator let* is also introduced to simplify error
handling. It's only defined locally in the mono phase for now.

See the added tests for the mono phase for examples on how this
transformation looks.