Global_checks: fix error in checking
This adds missing parentheses so that the global checks are applied
to all definitions (not just the standard library).

This bug was causing errors in the spec to be missed - they only
showed up because we use asli.exe to generate parts of the website
(and asli.exe didn't have this bug).