Extend transformation of assignment to bitslice
Extend the transformation with support for assignment to bitslice of a
field, for example field of record. The inner l-expression that yields
the record must be a variable or a field (recursive case).

This is part of an effort to simplify C backend by dropping the support
for assignment to bitslice.

Assignment to bitslice of array element is also a candidate for
extending this transformation (not part of this change).