Add transformation inserting calls to track valid bits
The transformation is optional: enabled by --calls-to-track-valid flag.

The transformation detects the statement where UNKNOWN expression is
assigned to a variable of bitvector type and inserts a procedure call
next to it. Intention of the called procedure is to track that the
variable got the UNKNOWN value and this way form/record validity mask
for the variable. The purpose of the mask is to exclude certain bits
during validation of the variable value, e.g. in the instruction set
simulator.

The only supported assignment statement for now is assignment to
bitslice. For such statement, the procedure call takes an index of the
variable, start bit of the slice and the width.

The start and the width of the bitslice must be expressions which are
safe to evaluate more than once. Otherwise, the transformation will not
fire.

The index of the variable is calculated from "track-valid" section of
the input .json file. This section is expected to list variable
identifiers. The index is the number of the variable identifier in this
list starting from 0.

Example:

    ASL code:

    var x : bits(3) {
        [0] f1
        [2:1] f2
    };

    x.f2 = UNKNOWN : bits(2);

    Input .json:

    {
        "track-valid": [
            "x"
        ]
    }

    Output C code:

    ASL_bits64_t x;
    x = ...
    ASL_fuzz(0LL, 1LL, 2LL);

---

Insert the transformation after the lowering transformation since it
does not handle hi:lo, single l-slices, or element *: slices.