Configuration constants: treat as variables in interpreter
This fixes a problem seen in ia-spec where use of configuration
constants whose initial value requires a call to a user-defined
function was causing xform_constprop to fail.

There is a remaining problem that configuration constants could
have dependencies on each other such as

    config X = TRUE;
    config Y = if X then 1 else 2;

This would evaluate fine if the initializers are executed in
the above order but it would still fail if they are executed
in reverse order.

The correct solution to this is not part of this commit and
requires that we either sort initializers according to
any dependencies between them or we use lazy evaluation to
dynamically discover the dependencies.
(Note that constants have the same problem as configuration
variables.)