Copyrights: check and update
This change adds missing copyright messages and corrects the years.

Copyright messages claiming Intel 2021 have been corrected
(we did not start changing things until 2022).

2024 is not included unless we actually changed the file in 2024.

Files that are heavily based on files that had Arm copyrights have had
an Arm copyright added.

In addition, an 'End' footer comment has been added where
it was missing for consistency with the rest of the codebase.