Verilog: split output into multiple files
This splits the generated code into three files in a similar way to the C backend

- <prefix>_types.svh
- <prefix>_vars.svh
- <prefix>_funs.sv

It does not split the functions into multiple files because that isn't helpful
for SystemVerilog.

It does not generate a <prefix>_exceptions.svh file because we don't yet support
exceptions in the Verilog backend.

Files do not `include other files because it is expected that this will be
done by the user of the generated code.