Tcheck: topologically sort definitions
In ASL, there is no need for definitions to come before uses but
this had not been fully implemented.

As we are restructuring the IA-spec into many separate files
(that are then read in alphabetical order), we are running into
this gap more and more often. That is, the typechecker was
complaining that there was no definition for X (because it had not
yet processed the definition).

This change performs a topological sort of the definitions so that
definitions come before uses - which solves the problem.

For testing purposes, we disable the sorting of definitions so that
we have more control/predictability of the input.