Fix handling of declarations with same name
This fixes a bug in 'reachable_decls' that was causing declarations to
be deleted if there were two declarations with the same name.
In particular, it was causing function prototypes (Decl_FunType and
Decl_ProcType) to be deleted - with a very subtle effect on the test
for reachable_decls.

The cause of the bug was incorrectly assuming that every
declaration has a unique name. This is not true for

- Decl_FunDefn and Decl_FunType
- Decl_ProcDefn and Decl_ProcType
- getter/setter pairs

[The bug became obvious with a change to topologically sort all
declarations prior to typechecking because it was causing function
definitions and getter/setters to be deleted!]