C backend: support accessing variables using a pointer
Global variable in expression or l-expression now can be accessed using
a C pointer.

When the generated C code is integrated into an instruction set
simulator, typically such a variable is not global but encapsulated
into a simulator-specific struct, e.g. processor state struct. Hence,
the variable needs to be referenced using a pointer to this struct.

The pointer can be global (assumed by this change) or passed along
with function arguments (not in this change).

The name of the global pointer is specified using --global-pointer
option.

The list of the global variables (identifiers) which should be accessed
using the pointer are specified in the new "pointer-accessed" group of
input .json files.

The change also includes new transformation, called delete_variables,
which deletes declarations of the global variables, specified in the
"pointer-accessed" group, because in the instruction set simulator the
variables should be declared elsewhere in the simulator-specific struct.