Syntax extension: array initializers
There was no syntax to create an array initializer in ASL.

Since [x,y] and {x,y} are already used, this uses the syntax

    array(x,y,z)  // 3 element array with elements x, y, and z

I am not sure whether this is the best syntax to use but it
will not be hard to change if we decide to change it.

At present, it is only supported when declaring global constants.

I'll propose this to Arm as an official extension.