Error reporting improvements
This changes typechecking so that, after the first typechecking failure,
the typechecker will skip to the next declaration and keep going.
This is more helpful when the code contains several type errors.

In order to implement this change, I also moved some of the typechecker's
exception declarations to the Error module.