Parser error reporting improvements
This uses the Menhir declarative parse error reporting mechanism.

This requires a file (libASL/asl_parser.messages) that provides
an error message for each error state in the parser.

Error states are identified by one of the example inputs that reaches
that state but, when reading / debugging the messages, you should look at what
grammar rules are involved.