Bugfix (xform_desugar)
This fixes a subtle bug in xform_desugar which was transforming

    add_bits_int{N}(x, y)

to

    add_bits{N}(x, y{bits(N)}[0 +: N])

which is not type correct.

The most obvious fix is to change the type in the slice to

    add_bits{N}(x, y{integer}[0 +: N])

but, since the function "cvt_bits_int" is not a builtin
function (i.e., it has a definition in the prelude), it is
better to think of this as a hardcoded inlining pass and
transform to

    add_bits{N}(x, cvt_int_bits{N}(y))

which is exactly what inlining would produce.