Fix opam install instructions
Although it is possible to use ASLi without installing alcotest
(one of the test dependencies), doing this has caused problems for
first time users.
So, to keep things simple, let's just install them all the time.
(The alcotest dependency is quite small in comparision to Z3 so
there is not much cost to doing this.)