runtime: fix print_int_dec/hex
print_int_dec was always printing in hex.
For values that fit in 64 bits, it is easy to do better.
For values that don't fit in 64-bits, we still (incorrectly) print in hex
but this is arguably a more useful way to print large values.

print_int_hex was always printing a 128 bit number
when something like "0x2a" was sufficient
and was printing negative numbers as 0xffffff2a instead of -0x2a