Lexer: extend bitliteral syntax
At the moment, ASL has just one syntax for bit literals: '1111'.

This adds three more forms:

- 32'hFFFF - hexadecimal
- 32'd1234 - decimal
- 32'b1111 - binary

This is an experimental extension. It has been proposed to the
ASL committee but it has not yet been officially accepted.
We should be careful when using this extension in specifications
because we may need to change it later.