Tests: delete buggy, low-value test
This deletes a fairly minor test that has been
causing problems because it was missing a needed
dependency.

An attempt to add the missing dependency seemed to
fix the problem and then it started failing again.
As with all missing dependency issues, whether it
works or not depends on the number of CPUs, how
fast they are, what has already been built, etc.
so this has been hard to fix.

What the test was checking was that, if you load
two 'project files', then they execute in the
expected order. This is a pretty minor test and
losing the test is not a big loss.