Makefile: cleanup opam installs
We learned the hard way that, when you uninstall a package,
opam does not remove all the files.
In particular, if you are using dune plugins and use dune install
and then you remove one of the plugins and dune install that,
then you can't run asli anymore.

So, we do the cleanup ourselves.