Internal extension: __assert e1 __in e2
This provides a way to perform an assertion in the middle
of an expression.

If condition e1 is TRUE, the result is e2.
If the condition e1 is FALSE, an assertion failure is reported.

This is intended for use in transformations and, in particular,
for performing runtime checks such as

    x[i]
-->
    __assert (i IN {0..31}) __in x[i]

Signed-off-by: Alastair Reid <alastair.reid@intel.com>