Xform_mono: generate function prototypes for instances
Previously, when we created a monomorphic instance, we
did not generate a function prototype. This caused problems
for the downstream backends that then had to generate C
function prototypes from function definitions which often resulted
in duplicate function prototypes.

This change generates the missing prototype which will allow
the backends to be simplified.

Signed-off-by: Alastair Reid <alastair.reid@intel.com>