Typechecker: drop constraints from mutable var decls
When constraints were added to the ASL language, they initially said that
in a mutable variable declaration like this

    var x = 0;

the type of x is "integer {0}".
This is because the type of "0" is "integer {0}".

But this makes no sense at all - the whole point of a *mutable* variable
is that you can assign it a value that is different from the
initial value.

This change corrects this weirdness by giving x the type "integer".