C23 runtime: fix integer promotion issues
Part of the fun of C's integer promotion rules is that,
if you add 2 8-bit unsigned values, the operation could
produce a different size of result because of the context
that the operation is performed in.

For example, if a and b are 8 bits, then

    uint64_t x = (uint64_t)(a + b);

could result in a 64-bit addition and, therefore, x[63:16]
may not be zero. (In particular, x[16] could be 1)

The way to guard against this context-dependent promotion
is to explicitly cast the result of every arithmetic operation
to the expected size.
For example, we calculate x like this

    uint64_t x = (uint64_t)(uint8_t)(x + y);

This commit inserts additional casts in every place
that we think an unintended promotion could occur.